Inf repel cheat ist now working on both editions and tested

+++ source/otherModifiers.h
@@ -3,7 +3,7 @@
 void initOtherModifiers(int edition);
 
 void setMoney(u32 count);
-void setRemainingRepel(u32 count);
+void setRemainingRepel(u8 count);
 void setRemainingOPower(u8 count);
 
 void setHatchingStepCounter(u16 steps);
